Merge in DNIND/real-time-edge-icc from ~NXF93252/real-time-edge-icc:master to master

* commit 'e2a7df5fa0c2149abd4e9bf7679623a71f43fe0f':
  Modify val type to unsigned int in icc_set_sgi
  Modify core_mask to unsigned int
  Modify memory size type to unsigned long

+++ inter-core-comm.c
@@ -201,9 +201,9 @@ void icc_block_free(unsigned long block)
 	blocks[block2index(block)] = 0;
 }
 
-void icc_set_sgi(int core_mask, unsigned int hw_irq)
+void icc_set_sgi(unsigned int core_mask, unsigned int hw_irq)
 {
-	unsigned long val;
+	unsigned int val;
 
 	if(hw_irq > 15) {
 		printf ("Interrupt id num: %u is not valid, SGI[0 - 15]\n", hw_irq);
